fix: setup deploy, typo, export types, edit feature
