+++ practice-one/src/constants/messages.ts
@@ -5,8 +5,8 @@ export const ERROR_MSG = {
   INVALID_EMAIL: 'Email must be a valid email address',
   INVALID_PASSWORD:
     'At least 6 characters, 1 uppercase, 1 digit and 1 special character',
-  INVALID_PHONENUMBER: 'Please enter a valid phone number format',
-  INVALID_ENROLLNUMBER: 'Enrollment number must be at least 8 characters',
+  INVALID_PHONE_NUMBER: 'Please enter a valid phone number format',
+  INVALID_ENROLL_NUMBER: 'Enrollment number must be at least 8 characters',
   WRONG_EMAIL_OR_PASSWORD: 'Wrong email or password',
 };
