+++ practice-one/src/constants/services.ts
@@ -1,5 +1,3 @@
-export const API_GATEWAY = 'http://localhost:3000';
-
 export const DATABASE_RESOURCES = {
   LOGIN: 'login',
   STUDENTS: 'students',
