--- practice-one/src/data/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "data",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "json-server": "^0.17.3",
-    "json-server-auth": "^2.1.0"
-  }
-}
