+++ practice-one/src/features/studentList/services/deleteStudent.ts
@@ -1,10 +1,10 @@
-import { API_GATEWAY, DATABASE_RESOURCES } from '@constants/services';
+import { DATABASE_RESOURCES } from '@constants/services';
 import { apiRequest } from '@services/apiRequest';
 
 export const removeStudent = async (studentId: string) => {
   try {
     return await apiRequest(
-      `${API_GATEWAY}/${DATABASE_RESOURCES.STUDENTS}/${studentId}`,
+      `${process.env.API_GATEWAY}/${DATABASE_RESOURCES.STUDENTS}/${studentId}`,
       'DELETE'
     );
   } catch (err) {
