+++ practice-one/src/services/apiRequest.ts
@@ -14,7 +14,7 @@ export const apiRequest = async <T>(
       body: JSON.stringify(body),
     });
     if (!response.ok) {
-      throw new Error(`${response.statusText}`);
+      throw new Error(`${response.status}`);
     }
 
     return await response.json();
