+++ practice-one/vite.config.ts
@@ -1,5 +1,5 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -16,4 +16,5 @@ export default defineConfig({
       '@utils': '/src/utils',
     },
   },
-})
+  define: { 'process.env': process.env },
+});
